+++ src/led.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2023 Ziga Miklosic
+// Copyright (c) 2025 Ziga Miklosic
 // All Rights Reserved
 // This software is under MIT licence (https://opensource.org/licenses/MIT)
 ////////////////////////////////////////////////////////////////////////////////
@@ -7,8 +7,8 @@
 * @brief    LED manipulations
 * @author   Ziga Miklosic
 * @email    ziga.miklosic@gmail.com
-* @date     08.11.2023
-* @version  V1.1.0
+* @date     01.02.2025
+* @version  V1.2.0
 */
 ////////////////////////////////////////////////////////////////////////////////
 /**
@@ -35,7 +35,7 @@
  *     Module version
  */
 #define LED_VER_MAJOR       ( 1 )
-#define LED_VER_MINOR       ( 1 )
+#define LED_VER_MINOR       ( 2 )
 #define LED_VER_DEVELOP     ( 0 )
 
 // Float definition
@@ -75,7 +75,8 @@ typedef enum
     {
         float32_t fade_in_time;     /**<Fade in time */
         float32_t fade_out_time;    /**<Fade out time */
-        float32_t max_duty;         /**<Maximum duty cycle */
+        float32_t max_duty;         /**<Maximum duty cycle in % */
+        float32_t min_duty;         /**<Minimum duty cycle in % */
     } led_fade_cfg_t;
 
 #endif
@@ -83,20 +84,26 @@ typedef enum
 ////////////////////////////////////////////////////////////////////////////////
 // Functions
 ////////////////////////////////////////////////////////////////////////////////
-led_status_t led_init           	(void);
+led_status_t led_init               (void);
 led_status_t led_deinit             (void);
-led_status_t led_is_init        	(bool * const p_is_init);
-led_status_t led_hndl           	(void);
-led_status_t led_set            	(const led_num_t num, const led_state_t state);
-led_status_t led_toggle         	(const led_num_t num);
-led_status_t led_blink          	(const led_num_t num, const float32_t on_time, const float32_t period, const led_blink_t blink);
-led_status_t led_get_active_time	(const led_num_t num, float32_t * const p_active_time);
-led_status_t led_is_idle        	(const led_num_t num, bool * const p_is_idle);
+led_status_t led_is_init            (bool * const p_is_init);
+led_status_t led_hndl               (void);
+led_status_t led_set                (const led_num_t num, const led_state_t state);
+led_status_t led_set_full           (const led_num_t num, const led_state_t state);
+led_status_t led_toggle             (const led_num_t num);
+led_status_t led_blink              (const led_num_t num, const float32_t on_time, const float32_t period, const led_blink_t blink);
+led_status_t led_get_active_time    (const led_num_t num, float32_t * const p_active_time);
+led_status_t led_is_idle            (const led_num_t num, bool * const p_is_idle);
+led_status_t led_is_on              (const led_num_t num, bool *const p_is_on);
+led_status_t led_set_on_brightness  (const led_num_t num, const float32_t duty_cycle);
+led_status_t led_set_off_brightness (const led_num_t num, const float32_t duty_cycle);
+led_status_t led_get_duty           (const led_num_t num, float32_t * const p_duty_cycle);
 
 #if ( 1 == LED_CFG_TIMER_USE_EN )
-    led_status_t led_set_smooth     (const led_num_t num, const led_state_t state);
-    led_status_t led_blink_smooth   (const led_num_t num, const float32_t on_time, const float32_t period, const led_blink_t blink);
-    led_status_t led_set_fade_cfg   (const led_num_t num, const led_fade_cfg_t * const p_fade_cfg);
+    led_status_t led_set_smooth             (const led_num_t num, const led_state_t state);
+    led_status_t led_blink_smooth           (const led_num_t num, const float32_t on_time, const float32_t period, const led_blink_t blink);
+    led_status_t led_set_fade_cfg           (const led_num_t num, const led_fade_cfg_t * const p_fade_cfg);
+    bool         led_is_in_smooth_blink_mode(const led_num_t num);
 #endif
 
 #endif // __LED_H_